pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "jobintel"
version = "0.1.0"
description = "US Job Market Intelligence Pipeline - Greenhouse & Lever data aggregation"
authors = [{name = "crussedev9"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.25.0",
    "pandas>=2.1.0",
    "pyarrow>=14.0.0",
    "pydantic>=2.5.0",
    "typer>=0.9.0",
    "tenacity>=8.2.0",
    "rich>=13.7.0",
    "loguru>=0.7.0",
    "pyyaml>=6.0.0",
    "python-dateutil>=2.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "black>=23.12.0",
    "ruff>=0.1.8",
    "mypy>=1.7.0",
]

[project.scripts]
jobintel = "jobintel.cli:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["src/jobintel/tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=jobintel --cov-report=html --cov-report=term"
